refactor ranking module
