+++ app/modules/auth/auth-types.ts
@@ -1,0 +1,7 @@
+export type User = {
+	id: string
+	username: string
+	isAdmin: number
+}
+
+// TODO: migrate to keratin authn -> create authn integration in this module
