+++ app/modules/posts/ranking/ranking-utils.ts
@@ -1,0 +1,25 @@
+import Immutable from 'immutable'
+import { type ImmutableReplyTree, type ReplyTree } from './ranking-types.ts'
+
+export function toImmutableReplyTree(replyTree: ReplyTree): ImmutableReplyTree {
+	return {
+		...replyTree,
+		replies: Immutable.List(replyTree.replies.map(toImmutableReplyTree)),
+	}
+}
+
+export function addReplyToReplyTree(
+	tree: ImmutableReplyTree,
+	reply: ImmutableReplyTree,
+): ImmutableReplyTree {
+	if (reply.post.parentId == tree.post.id) {
+		return {
+			...tree,
+			replies: tree.replies.insert(0, reply),
+		}
+	}
+	return {
+		...tree,
+		replies: tree.replies.map(child => addReplyToReplyTree(child, reply)),
+	}
+}
