+++ app/root.tsx
@@ -41,9 +41,9 @@ import {
 	DropdownMenuTrigger,
 } from './components/ui/dropdown-menu.tsx'
 import { href as iconsHref, Icon } from './components/ui/icon.tsx'
+import { type User } from './modules/auth/auth-types.ts'
 import { SITE_NAME } from './site.ts'
 import tailwindStyleSheetUrl from './styles/tailwind.css'
-import { type User } from './types/api-types.ts'
 import { getUserId, logout } from './utils/auth.server.ts'
 import { ClientHintCheck, getHints, useHints } from './utils/client-hints.tsx'
 import { csrf } from './utils/csrf.server.ts'
