+++ app/routes/polls.tsx
@@ -2,7 +2,7 @@ import { Link, useLoaderData } from '@remix-run/react'
 import { PollPostPreview } from '#app/components/building-blocks/poll-post-preview.tsx'
 import { Markdown } from '#app/components/markdown.tsx'
 import { db } from '#app/database/db.ts'
-import { getChronologicalPolls } from '#app/modules/posts/scoring/ranking-service.ts'
+import { getChronologicalPolls } from '#app/modules/posts/ranking/ranking-service.ts'
 
 export async function loader() {
 	const feed = await db.transaction().execute(async trx => {
